cast decompiler pointer artifacts through uintptr_t

+++ src/database/db_stringtable_load.cpp
@@ -3,7 +3,8 @@
 
 void __cdecl Load_ScriptStringCustom(uint16_t *var)
 {
-    *var = (uint16_t)varXAssetList->stringList.strings[*var];
+    // each strings[] slot holds a scr_string_t handle by this point
+    *var = (uint16_t)(uintptr_t)varXAssetList->stringList.strings[*var];
 }
 
 void __cdecl Mark_ScriptStringCustom(uint16_t *var)
